# docs/swagger.yaml
basePath: /v1
definitions:
...
  main.RegisterBookingSlotPayload:
    properties:
      days:
        items:
          type: string
        type: array
      end_period:
        type: string
      end_time:
        type: string
      start_period:
        type: string
      start_time:
        type: string
    type: object
  main.RegisterDisciplinePayload:
    properties:
      field:
        type: string
      subfield:
        type: string
    type: object
  main.RegisterEducationPayload:
    properties:
      degree:
        type: string
      field:
        type: string
      institute:
        type: string
      year_from:
        type: string
      year_to:
        type: string
    required:
    - degree
    - year_from
    type: object
...
  main.UpdateDisciplinePayload:
    properties:
      field:
        type: string
      subfield:
        type: string
    type: object
  main.UpdateEducationPayload:
    properties:
      degree:
        type: string
      field:
        type: string
      institute:
        type: string
      year_from:
        type: string
      year_to:
        type: string
    type: object
...
  store.Discipline:
    properties:
      field:
        type: string
      id:
        type: integer
      subfield:
        type: string
    type: object
  store.Education:
    properties:
      degree:
        type: string
      field:
        type: string
      id:
        type: integer
      institute:
        type: string
      userid:
        type: integer
      year_from:
        type: string
      year_to:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a Appointr server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Appointr API
  version: "1.0"
paths:
...
  /bookingslots/create:
    post:
      consumes:
      - application/json
      description: Create a new booking slot
      parameters:
      - description: Booking Slot
        in: body
        name: slot
        required: true
        schema:
          $ref: '#/definitions/main.RegisterBookingSlotPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Create a new booking slot
      tags:
      - booking
  /discipline:
    get:
      consumes:
      - application/json
      description: Fetches all discipline
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Discipline'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Fetches all discipline
      tags:
      - discipline
  /discipline/{disciplineField}:
    get:
      consumes:
      - application/json
      description: Fetches discipline by Field
      parameters:
      - description: Discipline Field
        in: path
        name: disciplineField
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Discipline'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Fetches discipline
      tags:
      - discipline
  /discipline/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a discipline field by ID
      parameters:
      - description: Discipline ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Deletes a discipline field
      tags:
      - discipline
    patch:
      consumes:
      - application/json
      description: Updates a discipline by ID
      parameters:
      - description: Discipline ID
        in: path
        name: id
        required: true
        type: integer
      - description: Discipline payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.UpdateDisciplinePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Discipline'
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Updates a discipline field
      tags:
      - discipline
  /discipline/create:
    post:
      consumes:
      - application/json
      description: creates a new discipline field
      parameters:
      - description: discipline
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.RegisterDisciplinePayload'
      produces:
      - application/json
      responses:
        "201":
          description: Discipline registered
          schema:
            type: integer
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Creates a new discipline
      tags:
      - discipline
  /education/{educationID}:
    delete:
      consumes:
      - application/json
      description: Delete education by id
      parameters:
      - description: Education ID
        in: path
        name: educationID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Delete education
      tags:
      - education
    patch:
      consumes:
      - application/json
      description: Update education
      parameters:
      - description: Education ID
        in: path
        name: educationID
        required: true
        type: integer
      - description: Education
        in: body
        name: education
        required: true
        schema:
          $ref: '#/definitions/main.UpdateEducationPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Education'
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Update education
      tags:
      - education
  /education/{educationid}:
    get:
      consumes:
      - application/json
      description: Get education by ID
      parameters:
      - description: Education ID
        in: path
        name: educationid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Education'
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get education by ID
      tags:
      - education
  /education/create:
    post:
      consumes:
      - application/json
      description: Create a new education
      parameters:
      - description: Education
        in: body
        name: education
        required: true
        schema:
          $ref: '#/definitions/main.RegisterEducationPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Create a new education
      tags:
      - education
  /education/u/{id}:
    get:
      consumes:
      - application/json
      description: Get all education by user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.Education'
            type: array
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get all education by user ID
      tags:
      - education
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"